# spec/swagger.yaml
definitions:
...
  handlers.ContainerGPURequest:
    properties:
      devices:
        items:
          type: string
        type: array
    type: object
  handlers.ContainerImageCapabilities:
    properties:
      dockerfile_build:
        type: boolean
      reason:
        type: string
    type: object
  handlers.ContainerImageCreateRequest:
    properties:
      build_options:
        additionalProperties: {}
        type: object
      dockerfile:
        type: string
      image_ref:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      source_type:
        type: string
    required:
    - image_ref
    - name
    - source_type
    type: object
  handlers.ContainerImageListPage:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.ContainerImageView'
        type: array
    type: object
  handlers.ContainerImageView:
    properties:
      build_options:
        additionalProperties: {}
        type: object
      builtin:
        type: boolean
      created_at:
        type: string
      digest:
        type: string
      dockerfile:
        type: string
      id:
        type: string
      image_ref:
        type: string
      last_build_error:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      owner_subject:
        type: string
      source_type:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.CreateContainerRequest:
    properties:
      gpu:
        $ref: '#/definitions/handlers.ContainerGPURequest'
      image:
        type: string
      local_workspace_path:
        type: string
      restore_data:
        type: boolean
      snapshotter:
        type: string
      workspace_backend:
        type: string
    type: object
  handlers.CreateContainerResponse:
    properties:
      cdi_devices:
        items:
          type: string
        type: array
      container_id:
        type: string
      container_path:
        type: string
      data_restored:
        type: boolean
      has_preserved_data:
        type: boolean
      image:
        type: string
      snapshotter:
        type: string
      started:
        type: boolean
      workspace_backend:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.GetContainerResponse:
    properties:
      cdi_devices:
        items:
          type: string
        type: array
      container_id:
        type: string
      container_path:
        type: string
      created_at:
        type: string
      has_preserved_data:
        type: boolean
      image:
        type: string
      legacy:
        type: boolean
      namespace:
        type: string
      status:
        type: string
      task_running:
        type: boolean
      updated_at:
        type: string
      workspace_backend:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/container:
    delete:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Export /data before deletion
        in: query
        name: preserve_data
        type: boolean
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete MCP container for bot
      tags:
      - containerd
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetContainerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get container info for bot
      tags:
      - containerd
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Create container payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateContainerRequest'
      responses:
        "200":
          description: SSE stream of container creation events
          schema:
            $ref: '#/definitions/handlers.CreateContainerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create and start MCP container for bot
      tags:
      - containerd
...
  /orchestration/container-images:
    get:
      description: Return the built-in images and tenant images available for env
        resources
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ContainerImageListPage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orchestration container images
      tags:
      - orchestration
    post:
      description: Register an existing image ref or Dockerfile build source for env
        resources
      parameters:
      - description: Container image payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.ContainerImageCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ContainerImageView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add an orchestration container image
      tags:
      - orchestration
  /orchestration/container-images/{id}:
    get:
      description: Return a single tenant image. Built-in images are returned by the
        list endpoint.
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ContainerImageView'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an orchestration container image
      tags:
      - orchestration
  /orchestration/container-images/capabilities:
    get:
      description: Return image build capabilities for the current runtime backend
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ContainerImageCapabilities'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get orchestration image capabilities
      tags:
      - orchestration
...
swagger: "2.0"